feat: 增加空文件检测/更新抓取逻辑

+++ index.ts
@@ -3,7 +3,7 @@ import chalk from 'chalk'
 import Axios from 'axios'
 import { sleep } from 'sleep'
 
-import { Headers, courseInfo } from './config'
+import { Headers, courseInfo, errList } from './config'
 
 declare interface CommonObj {
   [key: string]: string
@@ -79,20 +79,49 @@ function getSigleArticleContent(id: string) {
   })
 }
 
+// console.log(chalk.white.bgRed(`> Error: The config file is not exist! Please read README.md first! `))
+// process.exit(0)
 // note 主函数
 ;(async () => {
-  const articles = await getArticleList(courseInfo)
+  // 配置文件检测
+  if (!fs.existsSync('./config.ts')) {
+    console.log(chalk.white.bgRed(`> Error: The config file is not exist! Please read README.md first! `))
+    process.exit(1)
+  }
+
+  // 初始化输出目录
+  if (!fs.existsSync('./output/')) fs.mkdirSync('./output/')
+
+  // 初始化请求列表
+  let articles = await getArticleList(courseInfo)
+  if (errList.length) articles = errList
+  else articles = await getArticleList(courseInfo)
   // console.log(articles)
 
-  articles.forEach(async (info, index) => {
-    if (index > 0) sleep(5)
-    console.log(`request ${info.id} ---> ${info.title}\n`)
+  // 主逻辑
+  let curIndex = 0
+  const curErrList = []
+  while (curIndex < articles.length) {
+    const info = articles[curIndex]
 
+    console.log(`request ${info.id} ---> ${info.title}\n`)
     const data = await getSigleArticleContent(info.id)
 
     if (data.title) {
       fs.writeFileSync(`./output/${data.title.replace('/', '|')}.md`, data.result)
-      console.log(`write ${data.title} successfully \n`)
+      console.log(chalk.white.bgGreen(`write ${data.title} successfully \n`))
+    } else {
+      curErrList.push(info)
     }
-  })
+
+    sleep(5)
+    curIndex++
+  }
+
+  if (!curErrList.length) console.log(chalk.white.bgGreen('all successful'))
+  else {
+    console.log(chalk.yellow('\n> Error in get: '))
+    console.log(curErrList)
+    console.log(chalk.yellow('> You can copy the above array to config.ts.errList and try again'))
+  }
 })()
